+++ src/Exception/KeyNotExistsException.php
@@ -21,13 +21,38 @@ use Exception;
 class KeyNotExistsException extends Exception
 {
     /**
-     * Constructor
-     * @param string $message The string of the error message
-     * @param int $code The code of the error
-     * @param \Throwable $previous the previous exception
+     * @var string|null
      */
-    public function __construct($message = 'Array key does not exist', $code = 0, $previous = null)
+    protected $key;
+
+    /**
+     * Constructor
+     * @param string|null $message The string of the error message
+     * @param int $code The code of the error
+     * @param \Throwable|null $previous the previous exception
+     * @param string|null $key Name of the key that do not exist
+     * @uses $key
+     */
+    public function __construct($message = null, $code = 0, \Throwable $previous = null, $key = null)
     {
+        if (!$message) {
+            $message = 'Array key does not exist';
+            if ($key) {
+                $message = sprintf('Array key `%s` does not exist', $key);
+            }
+        }
         parent::__construct($message, $code, $previous);
+        $this->key = $key;
+    }
+
+    /**
+     * Gets the name of the key that do not exist
+     * @return string|null
+     * @since 1.2.11
+     * @uses $key
+     */
+    public function getKeyName()
+    {
+        return $this->key;
     }
 }